% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.BPF}
\alias{print.summary.BPF}
\title{Print a summary of a summary.BPF object}
\usage{
\method{print}{summary.BPF}(x, ..., margins = TRUE, digits = 2)
}
\arguments{
\item{x}{An \code{summary.BPF} class object.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{margins}{A \code{TRUE/FALSE} argument informing if the margins of the transition matrix should be displayed. Default, \code{TRUE}.}

\item{digits}{Integer indicating the number of decimal places to be shown. Default, 2.}
}
\value{
{No return value, called for side effects.}
}
\description{
Print method for \code{summary.BPFC} objects
}
